bot/pyproject.toml
[project]
name = "lms-telegram-bot"
version = "0.1.0"
description = "Telegram bot for LMS analytics"
requires-python = ">=3.12"
dependencies = ["python-dotenv>=1.0.0", "aiogram>=3.0.0", "httpx>=0.27.0"]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
